Ejemplo Python/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
grpcio
python-dotenv
